Convert subscriber notification signal into regular function call in views.
The previous method of using the post_save signal to notify subscribers meant that the function was called without the request available, creating an undocumented dependency on the django.contrib.sites app.  Moving the method call into the view enables the use of the get_current_site shortcut which removes the dependency.  It has the added benefit of ensuring that subscribers are only notified when a post is successfully added or edited in a thread, and not in any of the other possible scenarios where the post_save signal is sent.