Add custom signal for notifying topic subscribers, and a setting to disable notifications for edits.
Imho, notifications for post edits will be confusing to users the majority of the time, when they visit the thread and see no new posts.  The fraction of the time where they understand what has happened, the edit will be of no material value (typo fix, spelling/grammar) a lot of the time too.