Templates should use staticfiles, not STATIC_URL
As far as I can tell, it's generally preferred to use staticfiles and
{%static 'relative-filepath' %} instead of using STATIC_URL.
Additionally, as of my Django 1.8, I didn't get the static context
handler by default; this resulted in a good bit of digging around before
I found out what was causing STATIC_URL to be undefined.

The Django documentation on static file handling in templates doesn't
even mention using STATIC_URL anymore - it looks like Django 1.4 was the
last version that had documentation showing STATIC_URL being used this
way:
https://docs.djangoproject.com/en/1.4/howto/static-files/

Probably a good idea to update the templates to use staticfiles. This is
my attempt to do so.

	modified:   pybb/templates/pybb/_markitup.html
	modified:   pybb/templates/pybb/add_post.html
	modified:   pybb/templates/pybb/edit_post.html
	modified:   pybb/templates/pybb/post_form.html
	modified:   pybb/templates/pybb/post_template.html
	modified:   pybb/templates/pybb/topic.html