Allow ALL* tests to be runned with all DB
With a DB (like MySQL) which does not support microseconds, we force
a 1 second sleep for some operations to not fail some tests which
require a different datetime between two operations.
It add few seconds to run the test suite with mysql but it run all
tests (there were some tests which were skipped for mysql)

This commit allow to run test with sqlite, mysql or postgres
with tox for local testing.

Also fix travis issues:

* mysql has gone away
* only create DBs if required

Add methods `create_post` and `create_post_via_http` to simplify tests

Add a "problems" section in the doc about using DBs without microseconds
support to explain what are the side-effects.